feat(phase-19/02): expand rag-over-codebase into multi-file TypeScript project
